이미지 화질 개선, signup 페이지들 폴더로 분리, 뒤로가기 이미지에서 벡터로 교체
