+++ app/input-food/input-exer_1.tsx
@@ -51,11 +51,11 @@ export default function InputExer1() {
 
                 <View style={styles.pandaWrapper}>
 
-                    {/* 돋보기    */}
+                    {/* 돋보기 */}
                     <Image source={require('@/assets/images/readingglasses.png')} style={styles.readingglasses} />
                     
                     {/* 판다 */}
-                    <Image source={require('@/assets/images/sup_panda.png')} style={styles.panda} />
+                    <Image source={require('@/assets/images/surprised_panda_big.png')} style={styles.panda} />
                 </View>
 
                 { /* 텍스트 박스 */}
@@ -170,6 +170,7 @@ const styles = StyleSheet.create({
         justifyContent: 'center',
         gap: 20,
         marginTop: 20,
+        maxWidth: 300,
         paddingHorizontal: 10,
     },
     selectBox: {
