+++ app/input-food/input-exer_2.tsx
@@ -50,7 +50,7 @@ export default function InputExer2() {
                     <Image source={require('@/assets/images/readingglasses.png')} style={styles.pencil} />
 
                     {/* 판다 */}
-                    <Image source={require('@/assets/images/sup_panda.png')} style={styles.panda} />
+                    <Image source={require('@/assets/images/surprised_panda_big.png')} style={styles.panda} />
                 </View>
 
                 { /* 텍스트 박스 */}
