+++ app/input-food/input-food_5.tsx
@@ -52,7 +52,7 @@ export default function InputFood5() {
 
                 <View style={styles.pandaWrapper}>
                     {/* 판다 */}
-                    <Image source={require('@/assets/images/sup_panda.png')} style={styles.panda} />
+                    <Image source={require('@/assets/images/surprised_panda_big.png')} style={styles.panda} />
                 </View>
 
                 { /* 텍스트 박스 */}
